+++ src/routes/Router.tsx
@@ -9,6 +9,7 @@ import ChatRoom from '../pages/ChatRoom';
 import Footer from '../components/common/Footer';
 import SearchPage from '../pages/SearchPage';
 import OauthCallBack from '../pages/OauthCallBack';
+import PostModify from '../pages/PostModify';
 
 const Router = () => {
   return (
@@ -19,6 +20,7 @@ const Router = () => {
         <Route path="/" element={<Main />} />
         <Route path="/:id" element={<PostDetail />} />
         <Route path="/create" element={<PostCreate />} />
+        <Route path="/modify/:postId" element={<PostModify />} />
         <Route path="/mypage" element={<MyPage />} />
         <Route path="/chat" element={<ChatRoom />} />
         <Route path="/search/:word" element={<SearchPage />} />
